from master

+++ Progemon/src/logic/terrain/FightTerrain.java
@@ -1,6 +1,5 @@
 package logic.terrain;
 
-import java.awt.Graphics2D;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -15,7 +14,7 @@ import javafx.scene.image.Image;
 public class FightTerrain implements IRenderable {
 	
 	public final static int IMG_SIZE_X = 40, IMG_SIZE_Y = 40;
-	private int x, y;
+	private short x, y;
 	private boolean isShadowed, isCursor, isPathSign, isHighlight;
 	private TerrainType type;
 	private Image terrainImage = null;
